refactor(l1): split staffing demand types
